refactor(BA-6632): move image scanning out of the models layer
Move the container-registry scan/untag logic out of models/image/row.py
into the image repository layer so the models layer no longer imports the
container_registry package:

- scan_single_image / rescan_images (and the registry loading/filtering
  helpers) live on ImageDBSource.
- the registry untag external call lives on ImageRepository, which fetches
  the image + registry row first.
- cli/image_impl uses ImageDBSource for rescan and raises click errors in
  the purge untag path.
