+++ changes/12439.enhance.md
@@ -1,0 +1,1 @@
+Move container-registry image scanning out of `models/image/row.py` into the image repository layer (scan/rescan on the DB source, the registry untag on the repository) so the models layer no longer imports the `container_registry` package.
